修改long丢失精度的问题

+++ xxl-job-admin-pg/src/main/java/org/poem/controller/JobLogController.java
@@ -117,7 +117,7 @@ public class JobLogController {
 	}
 
 	@RequestMapping("/logDetailPage")
-	public String logDetailPage(int id, Model model){
+	public String logDetailPage(long id, Model model){
 
 		// base check
 		ReturnT<String> logStatue = ReturnT.SUCCESS;
